requirements.txt
paho-mqtt==1.6.1
RPi.GPIO==0.7.1
pi-rc522==2.3.0
PyYAML==6.0.2
pytest
pytest-mock